+++ libsrc/ffdec_lib/src/com/jpexs/video/FrameListener.java
@@ -1,0 +1,7 @@
+package com.jpexs.video;
+
+import java.awt.image.BufferedImage;
+
+public interface FrameListener {
+    public void newFrameRecieved(BufferedImage image);
+}