+++ src/com/sun/jna/platform/win32/WinReg.java
@@ -16,6 +16,7 @@
  */
 package com.sun.jna.platform.win32;
 
+import com.sun.jna.Native;
 import com.sun.jna.Pointer;
 import com.sun.jna.platform.win32.WinNT.HANDLE;
 import com.sun.jna.ptr.ByReference;
@@ -51,7 +52,7 @@ public interface WinReg extends StdCallLibrary {
         }
 
         public HKEYByReference(HKEY h) {
-            super(Pointer.SIZE);
+            super(Native.POINTER_SIZE);
             setValue(h);
         }